Split `SessionState` into its own module
